Refactors SampleModel and SampleModels
